backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.2.6
requests==2.32.3
python-dotenv==1.0.1
PyJWT==2.10.1
pytest==8.3.5
